backend/requirements.txt
# ============================================================
# Healthcare ML Project - Python Dependencies
# ============================================================
# Production-ready requirements for ML training pipeline
# Date: February 2026
# ============================================================

# Core Data Science Libraries
# ------------------------------------------------------------
numpy>=1.24.0              # Numerical computing
pandas>=2.0.0              # Data manipulation and analysis
scikit-learn>=1.3.0        # ML algorithms and preprocessing

# Machine Learning - XGBoost
# ------------------------------------------------------------
xgboost>=2.0.0             # Gradient boosting framework

# Model Persistence
# ------------------------------------------------------------
joblib>=1.3.0              # Model serialization

# Data Visualization
# ------------------------------------------------------------
matplotlib>=3.7.0          # Plotting library
seaborn>=0.12.0            # Statistical visualization

# Model Evaluation & Metrics
# ------------------------------------------------------------
# (included in scikit-learn)

# Logging and Utilities
# ------------------------------------------------------------
# (using Python standard library)

# Optional: Jupyter for notebooks
# ------------------------------------------------------------
jupyter>=1.0.0             # Jupyter notebook support
ipykernel>=6.0.0           # IPython kernel for Jupyter

# API Integration
# ------------------------------------------------------------
openai>=1.0.0              # OpenAI API client
python-dotenv>=1.0.0       # Load environment variables from .env

# Optional: Development tools
# ------------------------------------------------------------
# black>=23.0.0            # Code formatter
# flake8>=6.0.0            # Linter
# pytest>=7.0.0            # Testing framework
